+++ src/otf_site_data_rpc.cc
@@ -100,6 +100,9 @@ bool IsValidPermissionSetting(const std::string& permission,
   if (permission == "autoPictureInPicture") {
     return setting == "allow" || setting == "block";
   }
+  if (permission == "thirdPartyCookies") {
+    return setting == "allow" || setting == "block";
+  }
   if (permission == "images" || permission == "javascript") {
     return setting == "allow" || setting == "block";
   }